resource: move implemented resource algebras to resource dirs
