+++ source/vstd/resource/set.rs
@@ -67,9 +67,9 @@ use super::super::map::*;
 use super::super::map_lib::*;
 use super::super::modes::*;
 use super::super::prelude::*;
-use super::super::resource::Loc;
 use super::super::set::*;
 use super::super::set_lib::*;
+use super::Loc;
 use super::map::GhostMapAuth;
 use super::map::GhostPersistentPointsTo;
 use super::map::GhostPersistentSubmap;
