+++ source/vstd/tokens.rs
@@ -2,11 +2,6 @@ use super::multiset::*;
 use super::prelude::*;
 use core::marker::PhantomData;
 
-pub mod frac;
-pub mod map;
-pub mod seq;
-pub mod set;
-
 use verus as verus_;
 verus_! {
